Refactor services to unify portfolio summary and transaction services, update scheduling for data retrieval job, and enhance mobile totals display in instrument table
