+++ src/main/kotlin/ee/tenman/portfolio/controller/CalculatorController.kt
@@ -1,7 +1,8 @@
 package ee.tenman.portfolio.controller
 
 import ee.tenman.portfolio.service.CalculationResult
-import ee.tenman.portfolio.service.CalculatorService
+import ee.tenman.portfolio.service.CalculationService
+import ee.tenman.portfolio.service.SummaryService
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
@@ -11,8 +12,13 @@ import org.springframework.web.bind.annotation.RestController
 @RequestMapping("/api/calculator")
 @Validated
 class CalculatorController(
-  private val calculatorService: CalculatorService,
+  private val calculationService: CalculationService,
+  private val summaryService: SummaryService,
 ) {
   @GetMapping
-  fun calculate(): CalculationResult = calculatorService.getCalculationResult()
+  fun calculate(): CalculationResult {
+    val result = calculationService.getCalculationResult()
+    result.total = summaryService.getCurrentDaySummary().totalValue
+    return result
+  }
 }
