--- src/main/kotlin/ee/tenman/portfolio/service/CalculationResult.kt
@@ -1,12 +1,0 @@
-package ee.tenman.portfolio.service
-
-import ee.tenman.portfolio.service.xirr.Transaction
-import java.io.Serializable
-import java.math.BigDecimal
-
-data class CalculationResult(
-  var xirrs: List<Transaction> = mutableListOf(),
-  var median: Double = 0.0,
-  var average: Double = 0.0,
-  var total: BigDecimal = BigDecimal.ZERO,
-) : Serializable
