+++ src/main/kotlin/ee/tenman/portfolio/service/InstrumentService.kt
@@ -3,6 +3,7 @@ package ee.tenman.portfolio.service
 import ee.tenman.portfolio.configuration.RedisConfiguration.Companion.INSTRUMENT_CACHE
 import ee.tenman.portfolio.domain.Instrument
 import ee.tenman.portfolio.repository.InstrumentRepository
+import ee.tenman.portfolio.repository.PortfolioTransactionRepository
 import org.springframework.cache.annotation.CacheEvict
 import org.springframework.cache.annotation.Cacheable
 import org.springframework.cache.annotation.Caching
@@ -12,7 +13,7 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 class InstrumentService(
   private val instrumentRepository: InstrumentRepository,
-  private val portfolioTransactionService: PortfolioTransactionService,
+  private val portfolioTransactionRepository: PortfolioTransactionRepository,
   private val investmentMetricsService: InvestmentMetricsService,
   private val clock: java.time.Clock,
 ) {
@@ -50,12 +51,14 @@ class InstrumentService(
   @Transactional(readOnly = true)
   fun getAllInstruments(): List<Instrument> {
     val instruments = instrumentRepository.findAll()
-    val transactions = portfolioTransactionService.getAllTransactions().groupBy { it.instrument.id }
+    val allTransactions = portfolioTransactionRepository.findAllWithInstruments()
+    val transactions = allTransactions.groupBy { it.instrument.id }
     val calculationDate = java.time.LocalDate.now(clock)
 
     return instruments.map { instrument ->
       instrument.apply {
-        val metrics = investmentMetricsService.calculateInstrumentMetrics(this, transactions[id] ?: emptyList(), calculationDate)
+        val instrumentTransactions = transactions[id] ?: emptyList()
+        val metrics = investmentMetricsService.calculateInstrumentMetricsWithProfits(this, instrumentTransactions, calculationDate)
         totalInvestment = metrics.totalInvestment
         currentValue = metrics.currentValue
         profit = metrics.profit
